seperated save_measurements into its components and sorted them into the file system
