--- core/save_measurements.py
@@ -1,123 +1,0 @@
-#!/usr/bin/env python
-import os
-import glob
-import time
-import datetime
-import csv
-from zoneinfo import ZoneInfo
-from core.output.display_to_lcd import print_to_LCD
-from core.output.plot_to_commandline import plot_sensor_data
-
-device_name= "Peter" #TODO Create .env import
-measurement_interval=900 # Set the interval between measurements, in seconds: 900 = 15 min
-test_mode= False # To only print into command line, change to True
-
-# Sensor read functions
-
-def read_pressure():
-    # Pretend to read sensor data
-    return 1013.2
-
-def read_humidity():
-    return 45.6
-
-def read_temperature(decimals=1):
-    """Reads the temperature from a 1-wire device and returns it."""
-    device = glob.glob("/sys/bus/w1/devices/" + "28*")[0] + "/w1_slave"
-    try:
-        with open(device, "r") as f:
-            lines = f.readlines()
-        while lines[0].strip()[-3:] != "YES":
-            time.sleep(0.2)
-            with open(device, "r") as f:
-                lines = f.readlines() 
-        equals_pos = lines[1].find("t=")
-        if equals_pos != -1:
-            temp_string = lines[1][equals_pos+2:]
-            temp = round(float(temp_string) / 1000.0, decimals)
-            return temp
-    except IndexError:
-        print("Sensor not found or not connected properly.")
-        #TODO add error logging
-        return None
-    
-    
-    
-# Edit list to add/change sensor-readouts
-# First: label of sensor, second: function that reads sensor
-# Example: ("temperature", read_temperature)
-SENSORS = [
-    ("temperature", read_temperature),
-    ("pressure", read_pressure),
-    ("humidity", read_humidity)
-
-]
-
-
-def get_sensor_data(): # Calls sensor read functions from list SENSORS, creates dictionary with label of sensor and returned value
-    #TODO add timestamp here, but as first key in the dictionary
-    return {label: func() for label, func in SENSORS}    
-
-
-def write_data_to_csv(row_data, file_prefix=device_name, directory="data", max_rows=10000):
-    """Prints the measurement values to a CSV, adds header if necessary
-
-    :param row_data: (Dictionary of {Sensor_name, value})
-    :param file_prefix: (_type_, optional): identifier added to CSV file name. Defaults to device_name.
-    :param directory: (str, optional): Directory the files are saved in. Defaults to /data.
-    :param max_rows: (int, optional): max length of file. Defaults to 10000.
-    """
-    
-    timestamp = datetime.datetime.now(tz=ZoneInfo("Europe/Berlin"))
-    # Track how many rows have been written
-    # If no row or file numbering exist create them
-    if not hasattr(write_data_to_csv, "row_count"):
-        write_data_to_csv.row_count = 0
-        write_data_to_csv.file_index = 0
-
-    # Create folder if it doesn't exist
-    os.makedirs(directory, exist_ok=True)
-
-    # Rotate file if limit reached
-    if write_data_to_csv.row_count % max_rows == 0:
-        filename = f"{file_prefix}_{write_data_to_csv.file_index:03d}_{timestamp.isoformat()}.csv" 
-        write_data_to_csv.current_file = os.path.join(directory, filename)
-        write_data_to_csv.file_index += 1
-        write_data_to_csv.need_header = True
-    else:
-        write_data_to_csv.need_header = False
-
-    # Write data to the current file 
-    with open(write_data_to_csv.current_file, "a", newline="") as csv_file:
-        
-        fieldnames = ["timestamp"] + [label for label, _ in SENSORS] # Create the correct order of the column labels
-        row_data["timestamp"] = timestamp.strftime("%d/%m/%Y %H:%M:%S") #TODO Maybe change to .isoformat(): 2025-04-11T15:24:02
-        writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
-
-        if write_data_to_csv.need_header:
-            writer.writeheader()
-
-        writer.writerow(row_data)
-        write_data_to_csv.row_count += 1  
-        
-# Log data to csv indefinetly         
-def measurement_loop(test_mode):
-    
-    if test_mode: # Save data, print to console and to LCD-Display
-        
-        while True:        
-            sensor_data=get_sensor_data()
-            write_data_to_csv(sensor_data)
-            print(sensor_data)
-            print_to_LCD(measurement_values=sensor_data)
-            time.sleep(measurement_interval)
-            plot_sensor_data(write_data_to_csv.current_file)
-            
-    else:
-        while True:        
-            write_data_to_csv(get_sensor_data()) # Save data to csv
-            time.sleep(measurement_interval)
-        
-if __name__ == "__main__":
-        
-    measurement_loop(test_mode)
