+++ core/sensors/humid_demo.py
@@ -1,0 +1,5 @@
+def read_humidity():
+    
+    #pretend to measure humidity
+    
+    return 45.6
