+++ core/sensors/pres_demo.py
@@ -1,0 +1,3 @@
+def read_pressure():
+    # Pretend to read sensor data
+    return 1013.2
